% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{alpha_beta}
\alias{alpha_beta}
\alias{alpha_beta.wateres}
\title{Calculation of alpha and beta characteristics}
\usage{
alpha_beta(reser, alphas, max_beta, reliability, ...)

\method{alpha_beta}{wateres}(reser, alphas = seq(0, 1, 0.02),
  max_beta = 2, reliability = "max", ...)
}
\arguments{
\item{reser}{A \code{wateres} object.}

\item{alphas}{A vector of alpha values, i.e. coefficients by which mean annual flow will be multiplied. Usually the interval between 0 and 1 is used.}

\item{max_beta}{A maximum value of calculated beta to be considered, greater values will be ignored.}

\item{reliability}{A vector of reliability values passed to the \code{\link{sry.wateres}} function.}

\item{...}{Further arguments passed to the \code{\link{sry.wateres}} function (as \code{prob_type} or \code{upper_limit}).}
}
\value{
A \code{wateres_alpha_beta} object which is a data.table consisting of:
  \item{alpha}{level of development, given as the \code{alphas} argument}
  \item{beta}{ratio of storage representing the given reliability and volume of mean annual flow}
  \item{reliability}{given reliability values (may differ from reliabilities calculated by \code{\link{sry.wateres}})}
}
\description{
Calculates pairs of alpha (level of development) and beta (ratio of storage and volume of annual flow) characteristics of the reservoir
for given reliabilities.
}
\details{
An error occurs if the range given by \code{upper_limit} does not contain the given value of reliability. In that case, try to increase
  the \code{upper_limit} argument.
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, storage = 14.4e6, area = 754e3)
alpha_beta = alpha_beta(reser)
}
\seealso{
\code{\link{plot.wateres_alpha_beta}} for plotting the results, \code{\link{sry.wateres}} for calculation of a reservoir storage for the
  given yield and reliability
}
