% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootcast.R
\name{bootcast}
\alias{bootcast}
\title{Develop seasonal streamflow forecasts for an observed period ("hindcasts") using the k nearest neighbour (kNN) bootstrap in forecast mode.}
\usage{
bootcast(Q, H, start_yr, end_yr, k, d, sampling_mode, plot)
}
\arguments{
\item{Q}{time series object - seasonal streamflow rate or streamflow totals.}

\item{H}{integer giving the number of seasons in the forecast horizon (e.g., H = 3 means a three month forecast if the frequency of Q is 12).}

\item{start_yr}{start year (if omitted, forecast will be for one year)}

\item{end_yr}{end_year (if omitted, forecast will be for last year of input)}

\item{k}{integer. The k parameter of the kNN Bootstrap (i.e., number of nearest neighbors from which to sample). If left blank k = n ^ 0.5., where n is the number of years in the input data.}

\item{d}{integer. The d parameter of the kNN Bootstrap (i.e., number of previous time periods to inform the model). If left blank d = 1.}

\item{sampling_mode}{used to define period of observed streamflow data (Q) from which to sample the forecasts. "past" will use the period of data prior to the forecast start year. "all" will use all data within Q. "adapt" (the default) updates the sample each year to all years prior to forecast release.}

\item{plot}{logical. If TRUE (the default) the function will return a plot of the forecasts against observed flow during the period start_yr : end_yr.}
}
\value{
Returns the critical drawdown period of the reservoir, giving the shortest time from full to empty by default.
}
\description{
For seasonal streamflow "hindcasts" based solely on persistence.
}
\examples{
# prepare three month ahead forecasts of resX$Q_Mm3 for the period 1995 to 2000...
Q <- resX$Q_Mm3
Q_fcast <- bootcast(Q, H = 3, start_yr = 1995, end_yr = 2000)
}

