% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{calc_deficits}
\alias{calc_deficits}
\alias{calc_deficits.wateres_system}
\title{Calculation of system of reservoirs with respect to deficits}
\usage{
calc_deficits(system, yields)

\method{calc_deficits}{wateres_system}(system, yields)
}
\arguments{
\item{system}{A \code{wateres_system} object.}

\item{yields}{A vector of required fixed yield values in m3.s-1, its names have to correspond with the names of the reservoirs in the system.}
}
\value{
A list consisting of two items: \code{single} with the results for independently calculated reservoirs and \code{system} with the results
  for the system. Each of the items is a list of the \code{wateres_series} objects for individual reservoirs. The object contains the water
  balance variables returned by the \code{\link{calc_series}} functions. Moreover, \code{transfer} variable is added for the system results if has
  non-zero value at least in one time step.
}
\description{
Calculates time series of variables for reservoirs in the system in order to estimate influence of the system on deficit volumes.
Two variants are outputted: firstly, each reservoir is calculated independently of the system; secondly, water transfers decreasing
deficit volumes are added to the system.
}
\details{
The purpose of this system calculation is a rough estimation of deficit volumes elimination by the system. Therefore, water storage
is just redistributed within the system. Reservoir outflows are not considered as inflow to the reservoirs downwards; input time series of inflow
for individual reservoirs are used instead.

The water redistribution is carried out independently for the time steps when deficit in any of the reservoirs occurs. Future time steps are not
considered, i.e. it is possible that deficit decrease in a time step will cause deficit increase in the next time step.

For a given time step, the following algorithm is applied:

For each reservoir, the maximum water volume available to be transferred is determined (starting from top reservoirs and continuing with reservoirs
whose upper reservoirs have been already processed).
\itemize{
  \item{If sum of current reservoir storage and potential transfer is greater than zero, this sum is equal to the potential transfer from
        the current reservoir to the reservoir downwards.}
  \item{Otherwise, there is a deficit which can be satisfied by the potential transfer from upper reservoirs; then the exceeding water is equal to the potential transfer
        to the reservoir downwards. }}
Finally, volume resulting from all potential transfers is collected in the bottom reservoir. If it is greater than the bottom reservoir deficit,
the exceeding part is redistributed back to the corresponding source reservoirs. The only criterion of the redistribution is volume of the transfer, distance
between reservoirs is not taken into account. (The same redistribution method is applied when information about source reservoir is determined --
this happens if there is a potential transfer from another reservoir than bottom whose deficit is zero or has been completely satisfied.)

If provided in the reservoir object in the \code{system}, the time series of water balance variables are included the same way
as in the \code{\link{calc_series}} function. Other arguments of this function are set to their default values.

The system structure is checked by the \code{\link{check}} function before the calculation starts.
}
\examples{
period = seq(as.Date("2000-01-01"), by = "months", length.out = 24)
riv_data = data.frame(
    Q =  c(0.111, 0.339, 0.723, 0.165, 0.14, 0.088, 0.098, 0.052, 0.034, 0.022, 0.152, 0.162,
        0.156, 0.19, 0.259, 0.142, 0.075, 0.054, 0.118, 0.119, 0.267, 0.105, 0.194, 0.126),
    DTM = period)
riv = as.wateres(riv_data, 14.4e6, 754e3, id = "riv", down_id = "thar")
thar_data = data.frame(
    Q =  c(9.275, 32.586, 64.53, 16.702, 12.749, 9.646, 6.748, 6.645, 4.018, 3.523, 3.118, 4.009,
        7.137, 20.377, 47.467, 15.501, 8.199, 7.014, 7.086, 6.769, 9.038, 4.859, 12.006, 22.218),
    DTM = period)
thar = as.wateres(thar_data, 41.3e6, 2672e3, id = "thar")
sys = as.system(riv, thar)
resul = calc_deficits(sys, c(riv = 0.14, thar = 8))
}
\seealso{
\code{\link{calc_series}} for calculating individual reservoirs
}

