% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wateres.R
\name{calc_series}
\alias{calc_series}
\alias{calc_series.wateres}
\title{Calculation of reservoir time series}
\usage{
calc_series(reser, storage, yield, throw_exceed, initial_storage, initial_level,
  get_level)

\method{calc_series}{wateres}(reser, storage = attr(reser, "storage"), yield,
  throw_exceed = FALSE, initial_storage = storage, initial_level,
  get_level = FALSE)
}
\arguments{
\item{reser}{A \code{wateres} object.}

\item{storage}{A maximum reservoir storage in m3, if not given, take from the \code{reser} object.}

\item{yield}{A required yield in m3.s-1, either a value of fixed yield or a vector of the same length as the reservoir series.}

\item{throw_exceed}{Whether volume exceeding storage will be thrown or added to yield.}

\item{initial_storage}{A value of initial reservoir storage in m3. If not specified, the reservoir is considered to be full.}

\item{initial_level}{A value of initial water level in m.a.s.l. If specified and elevation-area-storage relationship is not provided within the
\code{reser} object, it will be ignored; otherwise the \code{initial_storage} argument will be ignored.}

\item{get_level}{Whether to obtain water level series for calculated storages. It is ignored if no elevation-area-storage relationship
is provided within the \code{reser} object.}
}
\value{
A \code{wateres_series} object which is a data table with water balance variables: storage (in m3), yield (in m3.s-1),
  precipitation, evaporation, withdrawal, deficits and transfer (in m3). The deficits represent the missing volume which would satisfy the remaining
  sum of yield and withdrawal demands. There is the water transfer only in case of non-zero values, resulting from calculations of a reservoir system.
  Positive values of transfer mean additional inflow whereas negative represent additional withdrawal.
  Additionally, water levels are included if the \code{get_level} argument is TRUE.
}
\description{
Calculates time series of water balance variables for the reservoir. If provided in the \code{reser} object, the precipitation, evaporation
or withdrawal variables are applied.
}
\details{
When calculating water balance, a simple explicit method is applied. Finally, the initial time step of storage is omitted
  to get a time series of the same length as for other variables.
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, storage = 14.4e6, area = 754e3)
resul = calc_series(reser, 14.4e6, 0.14)
}
\seealso{
\code{\link{plot.wateres_series}} for plotting the time series
}

