% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{calc_system}
\alias{calc_system}
\alias{calc_system.wateres_system}
\title{Calculation of system of reservoirs}
\usage{
calc_system(system, yields, initial_storages, types, yields_intercatch)

\method{calc_system}{wateres_system}(system, yields, initial_storages,
  types = c("single_plain", "system_plain"), yields_intercatch = FALSE)
}
\arguments{
\item{system}{A \code{wateres_system} object.}

\item{yields}{A vector of required fixed yield values in m3.s-1, its names have to correspond with the names of the reservoirs in the system. If values
for some reservoirs are not provided, they are taken from the \code{storage} property of that reservoirs (if the property is available).}

\item{initial_storages}{A vector of initial reservoir storages in m3 whose names correspond to the reservoirs names. If missing or NULL, value for each
reservoir is obtained from the \dQuote{storage_initial} property of reservoir; if that property is not available, the reservoir is considered
to be full initially.}

\item{types}{A vector of types of calculation whose valid values are \dQuote{single_plain}, \dQuote{system_plain}, \dQuote{single_transfer} and
\dQuote{system_transfer} (see details).}

\item{yields_intercatch}{Whether the vector of \code{yields} consists of values for intercatchment, i.e. whether total yields will be calculated as
a sum of yields from upstream reservoirs.}
}
\value{
A list consisting of items corresponding with the values of the \code{types} argument. Each of the items is a list of the \code{wateres_series}
objects for individual reservoirs. The object contains the water balance variables returned by the \code{\link{calc_series}} functions.
Moreover, \code{transfer} variable is added for the system results if has non-zero value at least in one time step.

If the yields are given as \code{yields_intercatch}, the list contains also the \code{yields} item with total yields calculated for the reservoirs.
}
\description{
Calculates time series of variables for reservoirs organized in a system. Four types of calculation are available, depending on whether
inflows from upstream reservoirs and water transfer between reservoirs are considered.
}
\details{
The types of calculation selected as the \code{types} argument are as follows:
\itemize{
\item{\code{single_plain} - reservoirs are calculated independently of the system.}
\item{\code{single_transfer} - as above, and water transfer is added to decrease deficit volumes; a fictional scenario for testing purposes.}
\item{\code{system_plain} - reservoirs are calculated within the system, i.e. reservoir inflow consists of yield of corresponding upstream
reservoirs and runoff from the intercatchment (derived from the original inflow series to reservoirs or given by the \dQuote{QI} column when
creating a reservoir by the \code{\link{as.wateres}} function).}
\item{\code{system_transfer} - as above, and water transfer is added.}}

The water transfer (redistribution) is carried out independently for the time steps when deficit in any of the reservoirs occurs. Future time steps are not
considered, i.e. it is possible that deficit decrease in a time step will cause deficit increase in the next time step.

For a given time step, the following algorithm is applied:

For each reservoir, the maximum water volume available to be transferred is determined (starting from top reservoirs and continuing with reservoirs
whose upper reservoirs have been already processed).
\itemize{
\item{If sum of current reservoir storage and potential transfer is greater than zero, this sum is equal to the potential transfer from
the current reservoir to the reservoir downwards.}
\item{Otherwise, there is a deficit which can be satisfied by the potential transfer from upper reservoirs; then the exceeding water is equal to the potential transfer
to the reservoir downwards. }}
Finally, volume resulting from all potential transfers is collected in the bottom reservoir. If it is greater than the bottom reservoir deficit,
the exceeding part is redistributed back to the corresponding source reservoirs. The only criterion of the redistribution is volume of the transfer, distance
between reservoirs is not taken into account. (The same redistribution method is applied when information about source reservoir is determined --
this happens if there is a potential transfer from another reservoir than bottom whose deficit is zero or has been completely satisfied.)

If provided in the reservoir object in the \code{system}, the time series of water balance variables are included the same way
as in the \code{\link{calc_series}} function. Other arguments of this function are set to their default values.

The system structure is checked by the \code{\link{check}} function before the calculation starts.
}
\examples{
period = seq(as.Date("2000-01-01"), by = "months", length.out = 24)
riv_data = data.frame(
    Q =  c(0.111, 0.339, 0.723, 0.165, 0.14, 0.088, 0.098, 0.052, 0.034, 0.022, 0.152, 0.162,
        0.156, 0.19, 0.259, 0.142, 0.075, 0.054, 0.118, 0.119, 0.267, 0.105, 0.194, 0.126),
    DTM = period)
riv = as.wateres(riv_data, 14.4e6, 754e3, id = "riv", down_id = "thar")
thar_data = data.frame(
    Q =  c(9.275, 32.586, 64.53, 16.702, 12.749, 9.646, 6.748, 6.645, 4.018, 3.523, 3.118, 4.009,
        7.137, 20.377, 47.467, 15.501, 8.199, 7.014, 7.086, 6.769, 9.038, 4.859, 12.006, 22.218),
    DTM = period)
thar = as.wateres(thar_data, 41.3e6, 2672e3, id = "thar")
sys = as.system(riv, thar)
resul = calc_system(sys, c(riv = 0.14, thar = 8))
}
\seealso{
\code{\link{calc_series}} for calculating individual reservoirs
}
