% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dirtyreps.R
\name{dirtyreps}
\alias{dirtyreps}
\title{Quick and dirty stochastic generation of seasonal streamflow replicates for a single site.}
\usage{
dirtyreps(Q, reps = 100, repyrs = length(Q)/frequency(Q), adjust = TRUE)
}
\arguments{
\item{Q}{time series object with seasonal resolution (e.g., frequency = 2, 3, 4, 6 or 12 for monthly data).}

\item{reps}{integer. The number of replicates to be generated.  The default is 100.}

\item{repyrs}{integer. The length of each replicate in years. The default is length(Q) / frequency(Q).}

\item{adjust}{logical. If TRUE (the default) the final output time series X will be coerced for 0 <= X <= 1.2*max(Q).}
}
\value{
Returns a multi time series object containing synthetic streamflow replicates.
}
\description{
Generates seasonal time series using a numerically-fitted PARMA(1,1) model. The function features automatic procedures for transforming seasonal sub-series to normal and deseasonalizing data.
}
\examples{
Q <- resX$Q_Mm3
replicates <- dirtyreps(Q)
mean(replicates); mean(Q)
sd(replicates); sd(Q)
layout(1:4)
plot(Q, main = "Recorded series", ylim = c(0,1500))
plot(replicates[,1], main = "Replicate # 1", ylim = c(0,1500))
plot(replicates[,20], main = "Replicate # 20", ylim = c(0,1500))
plot(replicates[,99], main = "Replicate # 99", ylim = c(0,1500))
}

