% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dp_hydro.R
\name{dp_hydro}
\alias{dp_hydro}
\title{Dynamic Programming for Hydropower Reservoir}
\usage{
dp_hydro(Q, capacity, capacity_live, surface_area, hydro_cap, head, qmax, depth,
  dep_vol_curve = "l", input_curve, S_disc = 1000, R_disc = 10,
  S_initial = 1, plot = TRUE)
}
\arguments{
\item{Q}{vector or time series object. Net inflows to the reservoir.}

\item{capacity}{numerical. The reservoir storage capacity (must be the same volumetric unit as Q).}

\item{capacity_live}{numerical. The reservoir live capacity (must be the same volumetric unit as Q).}

\item{surface_area}{numerical. The reservoir surface area}

\item{hydro_cap}{numerical. The hydropower plant electric capacity (MW)}

\item{head}{numerical. The maximum hydraulic head of the hydropower plant (m)}

\item{qmax}{numerical. The maximum flow into the hydropower plant}

\item{depth}{numerical. The reservoir maximum depth}

\item{dep_vol_curve}{string. The relationship between depth and volume of reservoir.}

\item{input_curve}{matrix of 3 columns: depth, area, volume}

\item{S_disc}{integer. Storage discretization--the number of equally-sized storage states. Default = 1000.}

\item{R_disc}{integer. Release discretization. Default = 10 divisions.}

\item{S_initial}{numeric. The initial storage as a ratio of capacity (0 <= S_initial <= 1). The default value is 1.}

\item{plot}{logical. If TRUE (the default) the storage behavior diagram and release time series are plotted.}
}
\value{
Returns the time series of optimal releases and, if requested, the reliability, resilience and vulnerability of the system.
}
\description{
Determines the optimal sequence of releases from the reservoir to minimise a penalty cost function based on water supply defict.
}
\examples{
\donttest{storage_cap <- 4 * mean(aggregate(ResX_inflow.ts)) # set storage ratio of 4 years
demand <- 0.8 * mean(ResX_inflow.ts) # set draft ratio of 0.8
optimal.releases <- dp(ResX_inflow.ts, capacity = storage_cap, target = demand)
}
}
\references{
Loucks, D.P., van Beek, E., Stedinger, J.R., Dijkman, J.P.M. and Villars, M.T. (2005) Water resources systems planning and management: An introduction to methods, models and applications. Unesco publishing, Paris, France.
}
\seealso{
\code{\link{sdp}} for Stochastic Dynamic Programming
}

