% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wateres.R
\name{fill_time}
\alias{fill_time}
\alias{fill_time.wateres}
\title{Reservoir filling time}
\usage{
fill_time(reser, yield, begins, samples)

\method{fill_time}{wateres}(reser, yield, begins = NULL, samples = 10)
}
\arguments{
\item{reser}{A \code{wateres} object.}

\item{yield}{A required yield.}

\item{begins}{A vector of time steps that represent begins of time series to be simulated.}

\item{samples}{A number of time steps of begins to be randomly sampled.}
}
\value{
A data table of filling times:
  \item{begin}{time steps when simulation of time series begins}
  \item{months}{number of months during that the reservoir changes from empty to full, or \code{NA} if the reservoir is not full and the end
    of the time series is reached}
}
\description{
Experimental calculation of filling time of the reservoir. A sample of time series is calculated based on the given beginning time steps (alternatively,
  these time steps are sampled randomly).
}
\details{
If provided, the \code{begins} argument is used to calculate the time series. Alternatively, the \code{samples} argument is applied.
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, storage = 4e5, area = 754e3)
fill = fill_time(reser, yield = 0.01)
}

