% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{GRanD}
\alias{GRanD}
\title{GRanD database}
\format{data frame object}
\source{
\url{http://atlas.gwsp.org}
}
\description{
GRanD database
}
\references{
Lehner, B., R-Liermann, C., Revenga, C., Vorosmarty, C., Fekete, B., Crouzet, P., Doll, P. et al.: High resolution mapping of the world's reservoirs and dams for sustainable river flow management. Frontiers in Ecology and the Environment. Source: GWSP Digital Water Atlas (2008). Map 81: GRanD Database (V1.0).
}

