% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{HollandCreek.ts}
\alias{HollandCreek.ts}
\title{Holland Creek streamflow 1961 - 2011 (monthly)}
\description{
Holland Creek streamflow 1961 - 2011 (monthly)
}
\author{
Sean Turner
}
\references{
Australian Bureau of Meteorology
}
\keyword{data_BOM}

