% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.wateres_alpha_beta}
\alias{plot.wateres_alpha_beta}
\title{Plot of alpha and beta characteristics}
\usage{
\method{plot}{wateres_alpha_beta}(x, filename = NULL, width = 8,
  height = 6, ...)
}
\arguments{
\item{x}{A \code{wateres_alpha_beta} object.}

\item{filename}{A file name where the plot will be saved. If not specified, the plot will be printed to the current device.}

\item{width}{Plot width in inches (or a unit specified by the \code{units} argument).}

\item{height}{Plot height in inches (or a unit specified by the \code{units} argument).}

\item{...}{Further arguments passed to the \code{\link[ggplot2:ggsave]{ggsave}} function saving the plot to a file.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots characteristics in a given \code{wateres_alpha_beta} object, by using the \code{ggplot2} package.
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, Vpot = 14.4, area = 0.754)
alpha_beta = alpha_beta(reser)
plot(alpha_beta)
}

