% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.wateres_prob_field}
\alias{plot.wateres_prob_field}
\title{Plot of probability field}
\usage{
\method{plot}{wateres_prob_field}(x, type = "storage", filename = NULL,
  width = 8, height = 6, ...)
}
\arguments{
\item{x}{A \code{wateres_prob_field} object.}

\item{type}{Type of values to be plotted (\dQuote{storage}, \dQuote{yield} or \dQuote{level}).}

\item{filename}{A file name where the plot will be saved. If not specified, the plot will be printed to the current device.}

\item{width}{Plot width in inches (or a unit specified by the \code{units} argument).}

\item{height}{Plot height in inches (or a unit specified by the \code{units} argument).}

\item{...}{Further arguments passed to the \code{\link[ggplot2:ggsave]{ggsave}} function saving the plot to a file.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots monthly values of storage, yield or level stored in a given \code{wateres_prob_field} object, by using the \code{ggplot2} package.
}
\details{
An error occurs if levels are required but they are not contained in the \code{wateres_prob_field} object.
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, Vpot = 14.4, area = 0.754)
prob_field = prob_field(reser, c(0.9, 0.95, 0.99), 0.14)
plot(prob_field, "storage")
}

