% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.wateres_series}
\alias{plot.wateres_series}
\title{Plot of reservoir time series}
\usage{
\method{plot}{wateres_series}(x, reser, type = "flow", begin = 1,
  end = nrow(x), filename = NULL, width = 8, height = 6, ...)
}
\arguments{
\item{x}{A \code{wateres_series} object with reservoir variables as returned by \code{\link{calc_series.wateres}}.}

\item{reser}{A corresponding \code{wateres} object whose series of number of minutes will be used.}

\item{type}{A type of variables to be plotted: \dQuote{storage}, \dQuote{level} or \dQuote{flow} which plots all of the remaining variables.}

\item{begin}{A time step to begin the plot.}

\item{end}{A time step to end the plot.}

\item{filename}{A file name where the plot will be saved. If not specified, the plot will be printed to the current device.}

\item{width}{Plot width in inches (or a unit specified by the \code{units} argument).}

\item{height}{Plot height in inches (or a unit specified by the \code{units} argument).}

\item{...}{Further arguments passed to the \code{\link[ggplot2:ggsave]{ggsave}} function saving the plot to a file.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots time series of chosen water balance variables for the reservoir, by using the \code{ggplot2} package.
}
\details{
Series of flows with all values equal to zero are ignored.

  An error occurs if length of the time series of results (\code{x}) does not equal to the length of the reservoir data (\code{reser}).
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, storage = 14.4e6, area = 754e3)
plot(calc_series(reser, 14.4e6, 0.14), reser)
}
\seealso{
\code{\link{calc_series.wateres}} used for calculation of reservoir time series
}
