% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{prob_field}
\alias{prob_field}
\alias{prob_field.wateres}
\title{Calculation of probability fields}
\usage{
prob_field(reser, probs, yield, storage, throw_exceed)

\method{prob_field}{wateres}(reser, probs, yield, storage = attr(reser,
  "Vpot"), throw_exceed = FALSE)
}
\arguments{
\item{reser}{A wateres object.}

\item{probs}{A vector of required probability values.}

\item{yield}{A value of yield to be used for calculation of storages in the reservoir.}

\item{storage}{A water reservoir storage value in millions of m3, the default value is equal to the potential volume of \code{reser}.}

\item{throw_exceed}{Whether volume exceeding storage will be thrown or added to yield (see also \code{\link{sry.wateres}}).}
}
\value{
A \code{wateres_prob_field} object which is a list consisting of:
  \item{quantiles}{data.table containing storage and yield values for months and given probabilities}
  \item{probs}{given probability values in percent as characters}
}
\description{
Calculates monthly values of storage and yield for given probabilities by using the \code{\link{quantile}} function with the default settings.
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, Vpot = 14.4)
prob_field = prob_field(reser, c(0.9, 0.95, 0.99), 0.14)
}

