\name{reservoir-package}
\alias{reservoir-package}
\alias{reservoir}
\docType{package}
\title{An R package for the analysis, design, and operation of water supply reservoirs}
\description{Measure single reservoir performance using resilience, reliability,
    and vulnerability metrics; develop storage-yield-reliability relationships; 
    determine no-fail Rippl storage with sequent peak analysis; optimize release
    decisions using determinisitc and stochastic dynamic programming; evaluate
    inflow characteristics.}
\details{
\tabular{ll}{
Package: \tab reservoir\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-17\cr
License: \tab GPL (>= 2)\cr
}
How to use the package overview...
}
\author{Who wrote it Maintainer: Who to complain to <yourfault@somewhere.net> Author}
\references{McMahon etc.}
\keyword{ package }
\examples{
# Example
}
