% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wateres.R
\name{resize_input}
\alias{resize_input}
\alias{resize_input.wateres}
\title{Reservoir input time series resize}
\usage{
resize_input(reser, begin, end)

\method{resize_input}{wateres}(reser, begin = 1, end = nrow(reser))
}
\arguments{
\item{reser}{A \code{wateres} object.}

\item{begin}{New time series begin entered as a value that can be converted to date or as a number representing index within the time series
(can be less than 1 for expanding before the current series). Dates cannot be used if the \code{wateres} object does not contain dates (e.g.
for hourly data).}

\item{end}{New time series end entered the same way as \code{begin}. The end value cannot be less than the begin value.}
}
\value{
A \code{wateres} object with the resized series.
}
\description{
Resizes input time series of the reservoirs. Both shortening and expanding are supported. In case of expanding, input variables for the new
time steps are set to zero while dates and number of minutes are correctly filled in.
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, storage = 4e5, area = 754e3)
reser = resize_input(reser, "2000-02-15", 20)
}

