% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ResX_inflow.ts}
\alias{ResX_inflow.ts}
\title{Reservoir X inflow time series 1964 - 2014 (monthly)}
\format{time series object}
\source{
\url{www.bom.gov.au/water/hrs/}
}
\description{
Reservoir X inflow time series 1964 - 2014 (monthly)
}
\examples{
plot(ResX_inflow.ts)
}
\author{
Sean Turner
}

