% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stoch_dyanamic_prog.R
\name{sdp}
\alias{sdp}
\title{Stochastic Dynamic Programming}
\usage{
sdp(Q, capacity, target, S_disc = 1000, R_disc = 10, Q_disc = c(0, 0.2375,
  0.475, 0.7125, 0.95, 1), loss_exp = 2, S_initial = 1, plot = TRUE,
  tol = 0.995, rep_rrv = FALSE)
}
\arguments{
\item{Q}{time series object. Net inflows to the reservoir.}

\item{capacity}{numerical. The reservoir storage capacity (must be same volumetric unit as Q and R).}

\item{target}{numerical. The target release constant.}

\item{S_disc}{integer. Storage discretization--the number of equally-sized storage states. Default = 1000.}

\item{R_disc}{integer. Release discretization--the allowed releases. Default = 10.}

\item{Q_disc}{vector. Inflow discretization bounding quantiles. Defaults to five inflow classes bounded by quantile vector c(0.0, 0.2375, 0.4750, 0.7125, 0.95, 1.0).}

\item{loss_exp}{numeric. The exponent of the penalty cost function--i.e., Cost[t] <- ((target - release[t]) / target) ^ **loss_exp**). Default value is 2.}

\item{S_initial}{numeric. The initial storage as a ratio of capacity (0 <= S_initial <= 1). The default value is 1.}

\item{plot}{logical. If TRUE (the default) the storage behavior diagram and release time series are plotted.}

\item{tol}{numerical. The tolerance for policy convergence. The default value is 0.995.}

\item{rep_rrv}{logical. If TRUE then reliability, resilience and vulnerability metrics are computed and returned.}
}
\value{
Returns the time series of optimal releases.
}
\description{
Derives the optimal release policy based on storage state, inflow class and within-year period.
}
\examples{
storage_cap <- 4 * mean(aggregate(HollandCreek.ts))
demand <- 0.8 * mean(HollandCreek.ts)
optimal.releases <- sdp(HollandCreek.ts, capacity = storage_cap, target = demand)
}
\references{
Loucks, D.P., van Beek, E., Stedinger, J.R., Dijkman, J.P.M. and Villars, M.T. (2005) Water resources systems planning and management: An introduction to methods, models and applications. Unesco publishing, Paris, France.

Gregory R. Warnes, Ben Bolker and Thomas Lumley (2014). gtools: Various R programming tools. R package version 3.4.1. http://CRAN.R-project.org/package=gtools
}
\seealso{
\code{\link{sdp}} for deterministic Dynamic Programming
}

