% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdp_simple_multi.R
\name{sdp_simple_multi}
\alias{sdp_simple_multi}
\title{Stochastic Dynamic Programming (Simplified, multi-objective)}
\usage{
sdp_simple_multi(Q, capacity, target, R_max, vol_targ = 0.75, weights = c(4,
  1, 1), S_disc = 1000, R_disc = 10, Q_disc = c(0, 0.2375, 0.475, 0.7125,
  0.95, 1), loss_exp = c(2, 2, 2), S_initial = 1, plot = TRUE,
  tol = 0.99, rep_rrv = FALSE)
}
\arguments{
\item{Q}{time series object. Net inflows to the reservoir.}

\item{capacity}{numerical. The reservoir storage capacity (must be the same volumetric unit as Q and the target release).}

\item{target}{numerical. The target release constant.}

\item{R_max}{numerical. The maximum controlled release.}

\item{vol_targ}{numerical. The target storage volume constant (as proportion of capacity).}

\item{weights}{vector of length 3 indicating weighting to be applied to release, spill and water level objectives respectively.}

\item{S_disc}{integer. Storage discretization--the number of equally-sized storage states. Default = 1000.}

\item{R_disc}{integer. Release discretization. Default = 10 divisions.}

\item{Q_disc}{vector. Inflow discretization bounding quantiles. Defaults to five inflow classes bounded by quantile vector c(0.0, 0.2375, 0.4750, 0.7125, 0.95, 1.0).}

\item{loss_exp}{vector of length 3 indicating the exponents on release, spill and water level deviations from target. Default exponents are c(2,2,2).}

\item{S_initial}{numeric. The initial storage as a ratio of capacity (0 <= S_initial <= 1). The default value is 1.}

\item{plot}{logical. If TRUE (the default) the storage behavior diagram and release time series are plotted.}

\item{tol}{numerical. The tolerance for policy convergence. The default value is 0.990.}

\item{rep_rrv}{logical. If TRUE then reliability, resilience and vulnerability metrics are computed and returned.}
}
\value{
Returns a list that includes: the optimal policy as an array of release decisions dependent on storage state, month/season, and current-period inflow class; the Bellman cost function based on storage state, month/season, and inflow class; the optimized release and storage time series through the training inflow data; the flow discretization (which is required if the output is to be implemented in the rrv function); and, if requested, the reliability, resilience, and vulnerability of the system under the optimized policy.
}
\description{
Derives the optimal release policy based on storage state and within-year period only. The objective is to minimise a penalty cost function based on water supply, spill, and water level.
}
\examples{
\donttest{storage_cap <- 4 * mean(aggregate(ResX_inflow.ts)) # set storage ratio of 4 years
demand <- 0.8 * mean(ResX_inflow.ts) # set draft ratio of 0.8
optimal.releases <- sdp_simple_multi(ResX_inflow.ts, capacity = storage_cap, target = demand)
}
}
\references{
Loucks, D.P., van Beek, E., Stedinger, J.R., Dijkman, J.P.M. and Villars, M.T. (2005) Water resources systems planning and management: An introduction to methods, models and applications. Unesco publishing, Paris, France.
}
\seealso{
\code{\link{sdp}} for deterministic Dynamic Programming
}

