% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{set_evaporation}
\alias{set_evaporation}
\alias{set_evaporation.wateres}
\title{Evaporation setting or calculation}
\usage{
set_evaporation(reser, values, altitude, plant_cover)

\method{set_evaporation}{wateres}(reser, values = NULL, altitude = NULL,
  plant_cover = NULL)
}
\arguments{
\item{reser}{A \code{wateres} object.}

\item{values}{A vector of evaporation values in mm, either monthly or daily values of length of reservoir time series, or 12 monthly values starting by January.}

\item{altitude}{Reservoir altitude (m.a.s.l.) used for calculation of monthly evaporation values according to the Czech Technical Standard
ČSN 75 2405, where evaporation is a function of altitude for range from 100 to 1200 m.}

\item{plant_cover}{Part of flooded area covered with plants, optional number between 0 and 0.75.}
}
\value{
A modified \code{wateres} object with evaporation time series added (denoted as \code{E}).
}
\description{
For monthly or daily data, sets or calculates time series of evaporation from the reservoir.
}
\details{
Evaporation is applied when calculating reservoir water balance. If no elevation-area-storage relationship is provided for the reservoir,
  evaporation only for flooded area related to the potential storage is assumed. Otherwise, evaporation is calculated for the area interpolated
  linearly by using the area-storage relationship (or area equal to the one of the limit value if storage fall out of the relationship limits).

  If the \code{plant_cover} argument is given, evaporation will be increased by plants transpiration as given by Vrána and Beran (1998). When the reservoir
  is getting empty, it is assumed that the plant cover area corresponds with the shallowest parts of the reservoir and therefore the coefficient increasing
  evaporation is adjusted accordingly.

  An error occurs if data in \code{reser} are not monthly or daily or if no dates are associated with daily data.
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, storage = 14.4e6, area = 754e3)
sry(reser, storage = 21e3, yield = 0.14)
reser = set_evaporation(reser, c(7, 14, 40, 62, 82, 96, 109, 102, 75, 48, 34, 13))
reser = set_evaporation(reser, altitude = 529)
resul = calc_series(reser, storage = 21e3, yield = 0.14)
}
\references{
ČSN 72 2405; Vrána, Karel, and Beran, Jan: Rybníky a účelové nádrže, ČVUT, Praha, 1998 (in Czech)
}

