% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{set_property}
\alias{set_property}
\alias{set_property.wateres}
\title{Reservoir property setting}
\usage{
set_property(reser, property_name, values)

\method{set_property}{wateres}(reser, property_name, values)
}
\arguments{
\item{reser}{A \code{wateres} object.}

\item{property_name}{One of \dQuote{storage} (maximum storage), \dQuote{storage_optim} (optimum storage), \dQuote{yield} (reservoir yield)
or \dQuote{yield_max} (maximum yield relevant if optimum, but not maximum storage is exceeded).}

\item{values}{One constant value or a vector of property values (in m3 for storages or m3.s-1 for yields), either monthly or daily of length of reservoir time series
(or plus one in case of storages), or 12 monthly values starting by January (for monthly or daily data only). For storages, the values relate to the beginning
of the particular month (e.g. the first of 12 values relates to the beginning of January, i.e. the end of December).}
}
\value{
A modified \code{wateres} object with the property added as its attribute.
}
\description{
Sets a value or time series of values of a reservoir property, i.e. characteristics which affect water balance calculation.
}
\details{
The reservoir properties are implemented as attributes of the \code{wateres} object.

If \dQuote{storage} is given, it replaces the value provided as the \code{storage} argument of the \code{as.wateres} function. However, \dQuote{yield}
given as a property is overriden by the \code{yield} argument of the \code{calc_series} function.
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, storage = 14.4e6, area = 754e3)
reser = set_property(reser, "storage", 1e6 * c(10, 10, 10, 14, 14, 14, 12, 12, 12, 6, 6, 6))
resul = calc_series(reser, yield = 0.17)
}
