% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment.R
\name{set_routing}
\alias{set_routing}
\alias{set_routing.catchment}
\title{Routing of catchment outflow setting}
\usage{
set_routing(reser, method, settings)

\method{set_routing}{catchment}(catchment, ...)
}
\arguments{
\item{catchment}{A \code{catchment} object.}

\item{...}{Routing method and settings as passed to the \code{\link{set_routing.wateres}} function.}
}
\value{
A \code{catchment} object with the routing settings.
}
\description{
Sets method and parameters used for routing for the catchment.
}
\details{
The catchment routing is implemented
as a routing for the outflow of the catchment, i.e. at its outlet (which is represented as a reservoir
with zero volume in the catchments).
}
\examples{
data_catch = data.frame(DTM = seq(as.Date("1982-11-01"), length.out = 7, by = "day"), PET = rep(0.5, 7), R = rep(24 * 3.6, 7), P = rep(1, 7))
res_data = data.frame(
    storage = c(1e7, 1e7, 1e7), area = c(1e4, 1e4, 1e4), part = c(0.25, 0.25, 0.5), branch_id = c("main", "lateral", "lateral"), id = c("M1", "L1", "L2"))
branches = list(main = list(down_id = NA), lateral = list(down_id = "main", connect_to_part = 0.8))
catch = as.catchment(id = "C1", down_id = "C2", data = data_catch, area = 100, res_data = res_data, branches = branches, main_branch = "main")
catch = set_routing(catch, "lag", list(lag_time = 2880))
}
