% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{set_routing}
\alias{set_routing}
\alias{set_routing.wateres}
\title{Routing of reservoir yield setting}
\usage{
set_routing(reser, method, settings)

\method{set_routing}{wateres}(reser, method = "none", settings = NULL)
}
\arguments{
\item{reser}{A \code{wateres} object.}

\item{method}{One of \dQuote{none} (no routing) or \dQuote{lag} (a simple shift of yield in time).}

\item{settings}{A list of routing parameters specific for each method: for \dQuote{lag} one parameter named \code{lag_time} defining time lag in minutes.}
}
\value{
A modified \code{wateres} object with the routing method and settings added as its attribute.
}
\description{
Sets method and parameters used for routing of reservoir yield.
}
\details{
The routing settings are implemented as attributes of the \code{wateres} object.
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, storage = 14.4e6, area = 754e3)
reser = set_routing(reser, "lag", list(lag_time = 88000))
resul = calc_series(reser, yield = 0.17)
}
