% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{set_wateruse}
\alias{set_wateruse}
\alias{set_wateruse.wateres}
\title{Water use setting}
\usage{
set_wateruse(reser, values)

\method{set_wateruse}{wateres}(reser, values)
}
\arguments{
\item{reser}{A \code{wateres} object.}

\item{values}{A vector of water use values in m3, either monthly or daily of length of reservoir time series, or 12 monthly values
starting by January (for monthly or daily time step of length one only), or one constant value. Positive values mean water release
to the reservoir, negative withdrawal from the reservoir.}
}
\value{
A modified \code{wateres} object with water use time series added (denoted as \code{W}).
}
\description{
Sets time series of water use for the reservoir.
}
\details{
Water use is applied when calculating reservoir water balance in two ways: positive water use (release) is always added to the storage,
whereas negative water use (withdrawal) is considered after the yield and evaporation demands are satisfied.
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, storage = 14.4e6, area = 754e3)
sry(reser, storage = 21e3, yield = 0.14)
reser = set_wateruse(reser, -1 * c(7, 14, 40, 62, 82, 96, 109, 102, 75, 48, 34, 13))
resul = calc_series(reser, storage = 21e3, yield = 0.14)
}
