% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{set_withdrawal}
\alias{set_withdrawal}
\alias{set_withdrawal.wateres}
\title{Withdrawal setting}
\usage{
set_withdrawal(reser, values)

\method{set_withdrawal}{wateres}(reser, values)
}
\arguments{
\item{reser}{A \code{wateres} object.}

\item{values}{A vector of withdrawal values in m3, either of length of reservoir time series, or 12 monthly values starting by January
(for monthly data only), or one constant value.}
}
\value{
A modified \code{wateres} object with withdrawal time series added (denoted as \code{W}).
}
\description{
Sets time series of withdrawal from the reservoir.
}
\details{
Withdrawal is applied when calculating reservoir water balance after the yield and evaporation demands are satisfied.
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, storage = 14.4e6, area = 754e3)
sry(reser, storage = 21e3, yield = 0.14)
reser = set_withdrawal(reser, c(7, 14, 40, 62, 82, 96, 109, 102, 75, 48, 34, 13))
sry(reser, storage = 21e3, yield = 0.14)
}

