% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wateres.R
\name{sry}
\alias{sry}
\alias{sry.wateres}
\title{Calculation of storage}
\usage{
sry(reser, reliability, yield, empirical_rel, upper_limit)

\method{sry}{wateres}(reser, reliability, yield, empirical_rel = TRUE,
  upper_limit = 1)
}
\arguments{
\item{reser}{A wateres object.}

\item{reliability}{A reliability value, cannot be less than zero or greater than maximum realiability value
(depending on data and usage of empirical reliability).}

\item{yield}{A required yield in m3.s-1, constant for all months.}

\item{empirical_rel}{Whether empirical probability (by (m - 0.3) / (n + 0.4) formula) will be used for calculation
of reliability. If enabled, maximum reliability value will be less than 1.}

\item{upper_limit}{An upper limit for optimization of storage given as multiple of potential volume of the reservoir.}
}
\value{
A list consisting of:
  \item{storage}{optimized storage value}
  \item{reliability}{reliability value for the optimized storage}
  \item{yield}{yield value equal to the \code{yield} argument}
}
\description{
Calculates water reservoir storage for given reliability and yield.
}
\details{
To optimize the value of storage, the \code{\link{optimize}} function is applied. If optimization fails, try
  to change its upper limit given as the \code{upper_limit} argument.
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, Vpot = 1)
sry(reser, reliab = 0.5, yield = 0.14)
}

