% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reservoir.R
\docType{package}
\name{wateres}
\alias{wateres}
\alias{wateres-package}
\title{Characteristics and Simulations for Water Reservoirs}
\description{
Tools to calculate characteristics of water reservoirs and to perform simulations for them.
}
\details{
A reservoir has to be created by \code{\link{as.wateres}} by using time series of dates
and corresponding flows. After that, characteristics for the reservoir can be calculated.
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, Vpot = 11.1)
summary(reser, Qn_coeff = c(0.1, 5, 0.1))
}

