% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{alpha_beta}
\alias{alpha_beta}
\alias{alpha_beta.wateres}
\title{Calculation of alpha and beta characteristics}
\usage{
alpha_beta(reser, yield_coeff, upper_limit)

\method{alpha_beta}{wateres}(reser, yield_coeff = c(0.1, 1.2, 0.05),
  upper_limit = 5)
}
\arguments{
\item{reser}{A wateres object.}

\item{yield_coeff}{A vector of alpha values, i.e. coefficients by which mean annual flow will be multiplied.}

\item{upper_limit}{An upper limit of storage (as multiple of the potential storage) for optimization as in the \code{\link{sry.wateres}} function.}
}
\value{
A \code{wateres_alpha_beta} object which is a data.table consisting of:
  \item{alpha}{level of development, given as the \code{yield_coeff} argument}
  \item{beta}{ratio of storage representing 100\% reliability and volume of yield}
}
\description{
Calculates pairs of alpha (level of development) and beta (ratio of storage and volume of yield) characteristics of the reservoir.
}
\details{
An error occurs if the range given by \code{upper_limit} does not contain value of 100\% reliability.
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, Vpot = 14.4, area = 0.754)
alpha_beta = alpha_beta(reser)
}
\seealso{
\code{\link{plot.wateres_alpha_beta}} for plotting the results
}

