% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment.R
\name{as.catchment}
\alias{as.catchment}
\title{Catchment with water reservoirs creation}
\usage{
as.catchment(id, down_id, data, area, res_data, branches, main_branch)
}
\arguments{
\item{branches}{A list of individual branches with reservoirs; list names correspond to reservoir IDs.
Each branch is represented by a list consisting of an ID of the downstream branch (\code{down_id}; NA for
the main branch in the catchment) and a point where the branch is connected to the downstream branch
(\code{connect_to_part}; not relevant for the main branch). The connection point is given as a catchment
area (relative to the whole catchment area) of the downstream branch before the junction with this branch.}

\item{main_branch}{An ID of the main branch, i.e. inflow from upstream catchments goes to this branch.}
}
\value{
A \code{catchment} object which is also of list class.
}
\description{
Creates an object of system of provided wateres reservoirs organized in a catchment.
}
\details{
An error occurs if there is a downstream branch which has not been provided in the \code{branches} list.
}
\examples{


}
