% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reservoir.R
\name{as.wateres}
\alias{as.wateres}
\title{Water reservoir creation}
\usage{
as.wateres(dframe, Vpot, observed = FALSE)
}
\arguments{
\item{dframe}{A name of file containing table with data (including header) or directly data frame or data table.
The data need to consist of monthly flows in m3.s-1 (\dQuote{Q} column) and dates (\dQuote{DTM} column).
Alternatively, this can be a Bilan object where the dates and modelled or observed runoffs are read from.
In that case, catchment area needs to be specified within the Bilan object.}

\item{Vpot}{Potential storage of the reservoir in millions of m3.}

\item{observed}{Only when Bilan object is used; whether to read observed runoffs from the object (otherwise modelled are read).}
}
\value{
A wateres object which is also of data.frame and data.table classes.
}
\description{
Creates a wateres object from provided time series.
}
\details{
An error occurs if \dQuote{Q} or \dQuote{DTM} column is missing or \code{dframe} is of another class
  than \code{data.frame} or \code{data.table}.
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, Vpot = 11.1)
}

