% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dp_hydro.R
\name{dp_hydro}
\alias{dp_hydro}
\title{Dynamic Programming for Hydropower Reservoir}
\usage{
dp_hydro(Q, capacity, capacity_live = capacity, surface_area, evap,
  installed_cap, head, qmax, max_depth, efficiency = 0.9, S_disc = 1000,
  R_disc = 10, S_initial = 1, plot = TRUE)
}
\arguments{
\item{Q}{time series object. Net inflows to the reservoir. Must be in volumetric units of Mm^3.}

\item{capacity}{numerical. The total reservoir storage capacity (including unusable "dead" storage). Must be in Mm^3.}

\item{capacity_live}{numerical. The volume of usable water in the reservoir ("live capacity" or "active storage"). capacity_live <= capacity. Default capacity_live = capacity. Must be in Mm^3.}

\item{surface_area}{numerical. The reservoir surface area at full capacity. Must be in square kilometers (km^2), or Mm^2.}

\item{evap}{vector or time series object. Pan evaporation, in units m. Varies with reservroi surface area..}

\item{installed_cap}{numerical. The hydropower plant electric capacity (MW).}

\item{head}{numerical. The maximum hydraulic head of the hydropower plant (m).}

\item{qmax}{numerical. The maximum flow into the hydropower plant.}

\item{max_depth}{numerical. The maximum water depth of the reservoir at the dam at maximum capacity. If omitted, the depth-storage-area relationship will be estimated from surface area and capacity only.}

\item{efficiency}{numerical. The hydropower plant efficiency. Default = 0.9.}

\item{S_disc}{integer. Storage discretization--the number of equally-sized storage states. Default = 1000.}

\item{R_disc}{integer. Release discretization. Default = 10 divisions.}

\item{S_initial}{numeric. The initial storage as a ratio of capacity (0 <= S_initial <= 1). The default value is 1.}

\item{plot}{logical. If TRUE (the default) the storage behavior diagram and release time series are plotted.}
}
\value{
Returns the time series of optimal releases and, if requested, the reliability, resilience and vulnerability of the system.
}
\description{
Determines the optimal sequence of releases from the reservoir to minimise a penalty cost function based on water supply defict.
}
\examples{
\donttest{storage_cap <- 4 * mean(aggregate(ResX_inflow.ts)) # set storage ratio of 4 years
demand <- 0.8 * mean(ResX_inflow.ts) # set draft ratio of 0.8
optimal.releases <- dp(ResX_inflow.ts, capacity = storage_cap, target = demand)
}
}
\references{
Loucks, D.P., van Beek, E., Stedinger, J.R., Dijkman, J.P.M. and Villars, M.T. (2005) Water resources systems planning and management: An introduction to methods, models and applications. Unesco publishing, Paris, France.
}
\seealso{
\code{\link{sdp}} for Stochastic Dynamic Programming
}

