% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurst_basic.R
\name{Hurst}
\alias{Hurst}
\title{Hurst coefficient estimation}
\usage{
Hurst(Q)
}
\arguments{
\item{Q}{vector or annualized time series object. Net inflows or streamflow totals.}
}
\value{
Returns an estimate of the Hurst coefficient, H (0.5 < H < 1).
}
\description{
Hurst coefficient estimation.
}
\examples{
Q_annual <- aggregate(resX$Q_Mm3) #convert monthly to annual data
Hurst(Q_annual)
}

