% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/res_rel_vul.R
\name{rrv}
\alias{rrv}
\title{Reliability, resilience, and vulnerability analysis for water supply reservoirs}
\usage{
rrv(X, target)
}
\arguments{
\item{X}{time series object representing the release time series of a reservoir.}

\item{target}{numerical constant (or a vector of length = length(X)). If omitted then the trigger constant will be < max(X).}
}
\value{
Returns reliability, resilience and vulnerability metrics based on supply deficits.
}
\description{
Computes time-based, annual, and volumetric reliability, as well as resilience and dimensionless vulnerability for a single reservoir.
}
\examples{
# Compare reliability, resilience and vulnerability for two operating policies (SOP and SDP).
}

