% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wateres.R
\name{summary.wateres}
\alias{summary.wateres}
\title{Water reservoir summary}
\usage{
\method{summary}{wateres}(object, ..., reliability = "max")
}
\arguments{
\item{object}{A \code{wateres} object.}

\item{...}{Further arguments passed to the \code{\link{sry.wateres}} function (as \code{storage}, \code{yield}, \code{prob_type} or \code{upper_limit}).}

\item{reliability}{A vector of reliability values passed to the \code{\link{sry.wateres}} function.}
}
\value{
A data table of reservoir characteristics:
  \item{storage}{reservoir storage in m3, given or the minimum storage calculated for given reliability and yield}
  \item{reliability}{given or calculated reliability}
  \item{yield}{the maximum yield (m3.s-1), given or calculated for given reliability and potential storage}
  \item{alpha}{level of development - ratio of yield to the mean annual flow}
  \item{m}{standardized net inflow - a measure of resilience calculated as (1 - alpha) / (standard deviation of annual flows / mean annual flow)}
  \item{resilience}{resilience calculated as number of continuous sequences of failures / total number of time steps with failures, NA for no failure}
  \item{vulnerability}{vulnerability (in m3) calculated as mean of monthly deficit volumes that represent maximum deficit of each failure period}
  \item{dimless_vulner}{dimensionless vulnerability, vulnerability value divided by yield value in volume units}
}
\description{
Calculates characteristics of the reservoir.
}
\details{
The maximum yield or the minimum storage is calculated by using the \code{\link{sry.wateres}} function for given storage or yield and reliability.

  To calculate reliability for given storage and yield without any optimization, provide all the three arguments: storage, reliability (will be ignored)
  and yield.

  An error occurs if the range given by \code{upper_limit} does not contain value of the given reliability or if an invalid reliability value or probability
  type is given.
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, storage = 14.4e6, area = 754e3)
summary(reser, reliability = 1)
summary(reser, reliability = 0.95)
}
\references{
Thomas A. McMahon, Adebayo J. Adeloye, Sen-Lin Zhou (2006): Understanding performance measures of reservoirs,
  Journal of Hydrology 324, 359–382.
}
\seealso{
\code{\link{sry.wateres}} used for optimization of the yield or storage for given reliability and for calculation of reliability
}

