% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wateres.R
\name{summary.wateres}
\alias{summary.wateres}
\title{Water reservoir summary}
\usage{
\method{summary}{wateres}(object, ..., Qn_coeff = c(0.1, 1.2, 0.05))
}
\arguments{
\item{object}{A wateres object.}

\item{...}{Further arguments are not used.}

\item{Qn_coeff}{Begin, end and step value of coefficients used to calculate designed yields to be tested.
The designed yields are equal to the mean annual flow multiplied by the coefficient.}
}
\value{
A vector of reservoir characteristics:
  \item{Vpot}{potential reservoir storage in millions of m3 (given as a parameter of \code{\link{as.wateres}})}
  \item{Qn_max}{the maximum yield (m3.s-1) for 100\% reliability for given potential storage}
  \item{alpha}{level of development - ratio Qn_max to the mean annual flow}
  \item{m}{resilience index - a measure of flow variability calculated as (1 - alpha) / (standard deviation of annual flows / mean annual flow)}
}
\description{
Calculates and shows characteristics of the reservoir.
}
\details{
An error occurs if potential storage \code{Vpot} is out of the range given by \code{Qn_coeff} so that
  value of Qn_max cannot be interpolated.
}
\examples{
reser = data.frame(
    Q = c(0.078, 0.065, 0.168, 0.711, 0.154, 0.107, 0.068, 0.057, 0.07, 0.485, 0.252, 0.236,
          0.498, 0.248, 0.547, 0.197, 0.283, 0.191, 0.104, 0.067, 0.046, 0.161, 0.16, 0.094),
    DTM = seq(as.Date("2000-01-01"), by = "months", length.out = 24))
reser = as.wateres(reser, Vpot = 14.4, area = 0.754)
summary(reser, Qn_coeff = c(0.1, 5, 0.1))
}

